Renamed ExportImageDialog files to ImageFileDialog
