+++ qtEditor/ImageFileDialog.h
